+++ asf_search/ASFSearchOptions/validators.py
@@ -2,7 +2,7 @@ import dateparser
 from datetime import datetime, timezone
 
 import requests
-from typing import Union, Tuple, TypeVar, Callable, List, Type, Sequence
+from typing import Dict, Union, Tuple, TypeVar, Callable, List, Type, Sequence
 
 import math
 from shapely import wkt, errors
@@ -22,7 +22,7 @@ def parse_string(value: str) -> str:
     except ValueError as exc: # If this happens, printing v's value would fail too...
         raise ValueError(f"Invalid string: Can't cast type '{type(value)}' to string.") from exc
     if len(value) == 0:
-        raise ValueError(f'Invalid string: Empty.')
+        raise ValueError('Invalid string: Empty.')
     return value
 
 
@@ -36,7 +36,7 @@ def parse_float(value: float) -> float:
         value = float(value)
     except ValueError as exc:
         raise ValueError(f'Invalid float: {value}') from exc
-    if math.isinf(value):
+    if math.isinf(value) or math.isnan(value):
         raise ValueError(f'Float values must be finite: got {value}')
     return value
 
@@ -109,9 +109,27 @@ def parse_list(value: Sequence, h) -> List:
     except ValueError as exc:
         raise ValueError(f'Invalid {h.__name__} list: {exc}') from exc
 
+def parse_cmr_keywords_list(value: Sequence[Dict]):
+    if not isinstance(value, Sequence):
+        value = [value]
+    
+    required_keys = ['key', 'fmt']
+
+    for idx, item in enumerate(value):
+        if not isinstance(item, dict):
+            raise ValueError(f"Expected item in umm search key list index {idx} to be dict, got value {item} of type {type(item)}")
+        for key in required_keys:
+            item_value = item.get(key)
+            if item_value is None:
+                raise ValueError(f"Expected key \"{key}\" in umm search key dict at index {idx}, got keys {item.keys()}")
+            elif not isinstance(item_value, str):
+                raise ValueError(f"Expected value at index {idx} for key \"{key}\" to be of type \"str\", got {type(item_value)} instead.")
+
+    return value
+
 # Parse and validate an iterable of strings: "foo,bar,baz"
 def parse_string_list(value: Sequence[str]) -> List[str]:
-    return parse_list(value, str)
+    return parse_list(value, parse_string)
 
 
 # Parse and validate an iterable of integers: "1,2,3"
@@ -199,6 +217,35 @@ def parse_wkt(value: str) -> str:
     except errors.WKTReadingError as exc:
         raise ValueError(f'Invalid wkt: {exc}') from exc
     return wkt.dumps(value)
+
+# Parse a CMR circle:
+#       [longitude, latitude, radius(meters)]
+def parse_circle(value: List[float]) -> List[float]:
+    value = parse_float_list(value)
+    if len(value) != 3:
+        raise ValueError(f'Invalid circle, must be 3 values (lat, long, radius). Got: {value}')
+    return value
+
+# Parse a CMR linestring:
+#       [longitude, latitude, longitude, latitude, ...]
+def parse_linestring(value: List[float]) -> List[float]:
+    value = parse_float_list(value)
+    if len(value) % 2 != 0:
+        raise ValueError(f'Invalid linestring, must be values of format (lat, long, lat, long, ...). Got: {value}')
+    return value
+
+def parse_point(value: List[float]) -> List[float]:
+    value = parse_float_list(value)
+    if len(value) != 2:
+        raise ValueError(f'Invalid point, must be values of format (lat, long). Got: {value}')
+    return value
+
+# Parse and validate a coordinate string
+def parse_coord_string(value: List):
+    value = parse_float_list(value)
+    if len(value) % 2 != 0:
+        raise ValueError(f'Invalid coordinate string, must be values of format (lat, long, lat, long, ...). Got: {value}')
+    return value
 
 # Take "requests.Session", or anything that subclasses it:
 def parse_session(session: Type[requests.Session]):